refactor: education items service into function
